Rename back to DoenetML.tsx
